Refactor: Updated module base packages and maven compiler plugin.
