+++ memo-api/src/main/java/com/github/elwyncrestha/api/user/service/UserService.java
@@ -1,0 +1,11 @@
+package com.github.elwyncrestha.api.user.service;
+
+import com.github.elwyncrestha.core.service.BaseService;
+import com.github.elwyncrestha.api.user.entity.User;
+
+/**
+ * @author Elvin Shrestha on 6/14/2020
+ */
+public interface UserService extends BaseService<User, Long> {
+
+}
