--- memo-api/src/main/java/com/github/elwyncrestha/digitalmemo/user/entity/User.java
@@ -1,67 +1,0 @@
-package com.github.elwyncrestha.digitalmemo.user.entity;
-
-import java.util.Collection;
-
-import javax.persistence.Entity;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.EqualsAndHashCode;
-import lombok.NoArgsConstructor;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-
-import com.github.elwyncrestha.digitalmemo.entity.BaseEntity;
-import com.github.elwyncrestha.digitalmemo.enums.Status;
-
-/**
- * @author Elvin Shrestha on 6/14/2020
- */
-@Entity
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@EqualsAndHashCode(callSuper = true)
-public class User extends BaseEntity<Long> implements UserDetails {
-
-    private String username;
-    private String password;
-    private String name;
-    private String email;
-    private Status status;
-
-    @Override
-    public Collection<? extends GrantedAuthority> getAuthorities() {
-        return null;
-    }
-
-    @Override
-    public String getPassword() {
-        return null;
-    }
-
-    @Override
-    public String getUsername() {
-        return username;
-    }
-
-    @Override
-    public boolean isAccountNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isAccountNonLocked() {
-        return true;
-    }
-
-    @Override
-    public boolean isCredentialsNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isEnabled() {
-        return status.equals(Status.ACTIVE);
-    }
-}
