--- memo-api/src/main/java/com/github/elwyncrestha/digitalmemo/user/service/UserService.java
@@ -1,11 +1,0 @@
-package com.github.elwyncrestha.digitalmemo.user.service;
-
-import com.github.elwyncrestha.digitalmemo.service.BaseService;
-import com.github.elwyncrestha.digitalmemo.user.entity.User;
-
-/**
- * @author Elvin Shrestha on 6/14/2020
- */
-public interface UserService extends BaseService<User, Long> {
-
-}
