+++ memo-core/src/main/java/com/github/elwyncrestha/core/enums/Status.java
@@ -1,0 +1,8 @@
+package com.github.elwyncrestha.core.enums;
+
+/**
+ * @author Elvin Shrestha on 6/14/2020
+ */
+public enum Status {
+    INACTIVE, ACTIVE, DELETED
+}
