+++ memo-core/src/main/java/com/github/elwyncrestha/core/exception/InvalidTokenException.java
@@ -1,0 +1,11 @@
+package com.github.elwyncrestha.core.exception;
+
+/**
+ * @author Elvin Shrestha on 6/14/2020
+ */
+public class InvalidTokenException extends RuntimeException {
+
+    public InvalidTokenException() {
+        super("Invalid Token");
+    }
+}
