--- memo-core/src/main/java/com/github/elwyncrestha/digitalmemo/constant/AppConstant.java
@@ -1,13 +1,0 @@
-package com.github.elwyncrestha.digitalmemo.constant;
-
-/**
- * @author Elvin Shrestha on 6/14/2020
- */
-public class AppConstant {
-
-    public static final String USERNAME = "username";
-    public static final String SPRING_MODEL = "spring";
-    public static final String SUCCESSFUL = "Successful";
-
-    private AppConstant() {}
-}
