--- memo-core/src/main/java/com/github/elwyncrestha/digitalmemo/entity/BaseEntity.java
@@ -1,49 +1,0 @@
-package com.github.elwyncrestha.digitalmemo.entity;
-
-import java.io.Serializable;
-import java.time.LocalDateTime;
-import javax.persistence.Column;
-import javax.persistence.EntityListeners;
-import javax.persistence.MappedSuperclass;
-import javax.persistence.Version;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import lombok.AccessLevel;
-import lombok.EqualsAndHashCode;
-import lombok.Getter;
-import lombok.Setter;
-import org.springframework.data.annotation.CreatedBy;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedBy;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.jpa.domain.AbstractPersistable;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-/**
- * @author Elvin Shrestha on 6/14/2020
- */
-@Getter(AccessLevel.PROTECTED)
-@Setter(AccessLevel.PROTECTED)
-@EqualsAndHashCode(callSuper = true)
-@MappedSuperclass
-@EntityListeners(AuditingEntityListener.class)
-@JsonIgnoreProperties({"hibernateLazyInitializer", "handler"})
-public class BaseEntity<PK extends Serializable> extends AbstractPersistable<PK> {
-
-    @CreatedDate
-    @Column(nullable = false, updatable = false)
-    private LocalDateTime createdAt = LocalDateTime.now();
-
-    @LastModifiedDate
-    @Column(nullable = false, updatable = false)
-    private LocalDateTime lastModifiedAt = LocalDateTime.now();
-
-    @CreatedBy
-    private Long createdBy;
-
-    @LastModifiedBy
-    private Long lastModifiedBy;
-
-    @Version
-    private int version;
-}
