--- memo-core/src/main/java/com/github/elwyncrestha/digitalmemo/enums/Status.java
@@ -1,8 +1,0 @@
-package com.github.elwyncrestha.digitalmemo.enums;
-
-/**
- * @author Elvin Shrestha on 6/14/2020
- */
-public enum Status {
-    INACTIVE, ACTIVE, DELETED
-}
