--- memo-core/src/main/java/com/github/elwyncrestha/digitalmemo/exception/InvalidTokenException.java
@@ -1,11 +1,0 @@
-package com.github.elwyncrestha.digitalmemo.exception;
-
-/**
- * @author Elvin Shrestha on 6/14/2020
- */
-public class InvalidTokenException extends RuntimeException {
-
-    public InvalidTokenException() {
-        super("Invalid Token");
-    }
-}
