--- memo-web/src/test/java/com/github/elwyncrestha/digitalmemo/DigitalMemoApplicationTests.java
@@ -1,11 +1,0 @@
-package com.github.elwyncrestha.digitalmemo;
-
-import org.junit.jupiter.api.Test;
-
-class DigitalMemoApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
