+++ memo-web/src/test/java/com/github/elwyncrestha/web/DigitalMemoApplicationTests.java
@@ -1,0 +1,11 @@
+package com.github.elwyncrestha.web;
+
+import org.junit.jupiter.api.Test;
+
+class DigitalMemoApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
